Automatizované Testování, RBAC

+++ backend/src/controllers/spell.controller.ts
@@ -1,5 +1,6 @@
 import { Request, Response } from 'express';
 import { PrismaClient } from '@prisma/client';
+import { AuthRequest } from '../middlewares/auth.middleware'
 
 const prisma = new PrismaClient();
 
@@ -37,24 +38,39 @@ export const getSpellById = async (req: Request, res: Response) => {
     }
 };
 
-export const createSpell = async (req: Request, res: Response) => {
+export const createSpell = async (req: AuthRequest, res: Response) => {
     try {
+
+        if (req.user?.role === 'USER') {
+            return res.status(403).json({ error: 'Pro tvorbu homebrew potřebuješ roli HOMEBREW nebo ADMIN.' });
+        }
+
         const { name, source, level, school, data } = req.body;
         const newSpell = await prisma.spell.create({
-            data: { name, source, level, school: school || 'Unknown', data: data || {} }
+            data: {
+                name, source, level, school: school || 'Unknown', data: data || {},
+                authorId: req.user?.userId
+            }
         });
         res.status(201).json({ message: 'Spell created successfully', data: newSpell });
-    } catch (error) {
-        res.status(500).json({ error: 'Failed to create spell' });
-    }
+    } catch (error) { res.status(500).json({ error: 'Failed to create spell' }); }
 };
 
-export const deleteSpell = async (req: Request, res: Response) => {
+export const deleteSpell = async (req: AuthRequest, res: Response) => {
     try {
         const { id } = req.params;
+        const spell = await prisma.spell.findUnique({ where: { id } });
+
+        if (!spell) return res.status(404).json({ error: 'Spell not found' });
+
+        const isOwner = spell.authorId === req.user?.userId;
+        const isAdmin = req.user?.role === 'ADMIN';
+
+        if (!isOwner && !isAdmin) {
+            return res.status(403).json({ error: 'Můžeš mazat jen svá vlastní kouzla (pokud nejsi admin).' });
+        }
+
         await prisma.spell.delete({ where: { id } });
         res.json({ message: 'Spell deleted successfully' });
-    } catch (error) {
-        res.status(500).json({ error: 'Failed to delete spell' });
-    }
+    } catch (error) { res.status(500).json({ error: 'Failed to delete spell' }); }
 };
